Expand filepath to ensure files are loaded only once
