+++ lib/base/rcov_analyzer.rb
