+++ spec/spec_helper.rb
@@ -3,5 +3,5 @@ require 'spec/autorun'
 require 'date'
 require 'construct'
 
-require File.join(File.dirname(__FILE__), '/../lib/metric_fu.rb')
+require File.expand_path File.join(File.dirname(__FILE__), '/../lib/metric_fu.rb')
 include MetricFu
